Refactor project Algoloop.Model
