+++ Algoloop.Model/AccountModel.cs
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-using Algoloop.Support;
+using Algoloop.Model.Converter;
 using System;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
