+++ Algoloop.Model/BalanceModel.cs
