+++ Algoloop.Model/Converter/ProviderNameConverter.cs
@@ -1,5 +1,5 @@
-﻿/*
- * Copyright 2018 Capnode AB
+/*
+ * Copyright 2019 Capnode AB
  * 
  * Licensed under the Apache License, Version 2.0 (the "License"); 
  * you may not use this file except in compliance with the License.
@@ -12,23 +12,14 @@
  * limitations under the License.
 */
 
-using Algoloop.Model;
-using GalaSoft.MvvmLight.Ioc;
+using QuantConnect;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
 
-namespace Algoloop.Support
+namespace Algoloop.Model.Converter
 {
-    public class AccountNameConverter : TypeConverter
+    internal class ProviderNameConverter : TypeConverter
     {
-        private readonly MarketsModel _markets;
-
-        public AccountNameConverter()
-        {
-            _markets = SimpleIoc.Default.GetInstance<MarketsModel>();
-        }
-
         public override bool GetStandardValuesSupported(ITypeDescriptorContext context)
         {
             // true means show a combobox
@@ -43,15 +34,16 @@ namespace Algoloop.Support
 
         public override StandardValuesCollection GetStandardValues(ITypeDescriptorContext context)
         {
-            // Request list of accounts
-            IReadOnlyList<AccountModel> accounts = _markets.GetAccounts();
-            List<string> list = accounts
-                .Select(m => m.DisplayName)
-                .ToList();
+            int code = 1;
+            string name;
+            List<string> names = new List<string>();
+            while ((name = Market.Decode(code++)) != null)
+            {
+                names.Add(name);
+            }
 
-            list.Sort();
-
-            return new StandardValuesCollection(list);
+            names.Sort();
+            return new StandardValuesCollection(names);
         }
     }
 }
