+++ Algoloop.Model/ListModel.cs
