+++ Algoloop.Model/MarketsModel.cs
