+++ Algoloop.Model/ModelBase.cs
