+++ Algoloop.Model/ParameterModel.cs
