+++ Algoloop.Model/PositionModel.cs
