+++ Algoloop.Model/SettingModel.cs
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-using Algoloop.Support;
+using Algoloop.Model.Converter;
 using System;
 using System.ComponentModel;
 using System.IO;
