+++ Algoloop.Model/StrategiesModel.cs
