+++ Algoloop.Model/StrategyModel.cs
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-using Algoloop.Support;
+using Algoloop.Model.Converter;
 using QuantConnect;
 using System;
 using System.Collections.ObjectModel;
