+++ Algoloop.Tests/Lean/LeanLauncherTests.cs
@@ -13,7 +13,6 @@
  */
 
 using Algoloop.Model;
-using Algoloop.Support;
 using Algoloop.Wpf.Lean;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using QuantConnect;
