+++ Algoloop.Wpf/Algoloop.Wpf.csproj
@@ -186,7 +186,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Algoloop.Brokerages\Algoloop.Brokerages.csproj" />
-    <ProjectReference Include="..\Algoloop.Core\Algoloop.Model.csproj" />
     <ProjectReference Include="..\Algoloop.Model\Algoloop.Model.csproj" />
     <ProjectReference Include="..\Algorithm.CSharp\QuantConnect.Algorithm.CSharp.csproj" />
     <ProjectReference Include="..\Algorithm.Framework\QuantConnect.Algorithm.Framework.csproj" />
