+++ Algoloop.Wpf/Provider/ProviderBase.cs
@@ -13,7 +13,6 @@
  */
 
 using Algoloop.Model;
-using Algoloop.Support;
 using Algoloop.Wpf.Common;
 using Newtonsoft.Json;
 using QuantConnect;
