+++ Algoloop.Wpf/ViewModel/MarketsViewModel.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Copyright 2018 Capnode AB
  * 
  * Licensed under the Apache License, Version 2.0 (the "License"); 
@@ -13,7 +13,6 @@
  */
 
 using Algoloop.Model;
-using Algoloop.Support;
 using Algoloop.Wpf.ViewSupport;
 using GalaSoft.MvvmLight.Command;
 using Newtonsoft.Json;
